refactor: rename solution 2 file
