+++ packages/date-picker/src/Filter/index.ts
@@ -1,2 +1,2 @@
 export * from "./FilterSolution2Ref"
-export * from "./FilterSolution2Select"
+export * from "./FilterSolution2"
